chore: relocate middleware, extract constants
Move security-headers and index-html middleware into src/server/modules/middleware to match the module-based layout, and extract the CSP nonce key and placeholder regex into a shared middleware.constants.ts so both middlewares reference the same source of truth.
